+++ src/components/app/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Switch, Route, BrowserRouter } from "react-router-dom";
-import Home from "../Home";
-import Waiting from "../Waiting";
-import InGame from "../Ingame";
+import Home from "../../pages/Home";
+import Waiting from "../../pages/Waiting";
+import InGame from "../../pages/Ingame";
 
 export default function App() {
   return (
