+++ src/pages/Waiting/index.js
